rss/pyproject.toml
[project]
name = "rss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "discord-py>=2.4.0",
    "feedparser>=6.0.11",
    "langchain-community>=0.3.16",
    "langchain-deepseek-official>=0.1.0",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.3",
    "ollama>=0.4.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "transformers>=4.48.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]


# uvの記述は省略 ...

[tool.ruff]
# 1行の最大文字数
line-length = 88

[tool.ruff.lint]
# チェックするエラーの種類
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
# 除外するエラーの種類
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904", # raise ... from ... になっていない場合のエラーを無視
]

# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

# isort の設定
[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.lint.pyupgrade]
# Python3.8互換のための設定
keep-runtime-typing = true
